requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.11
scikit-learn
dill
mlflow
pyaml
fastapi
uvicorn
dagshub
python-multipart

##-e .